R.js optimizer usage in Web Publishing Pipeline.
Things to improve: optimizer takes files directly from the project
directory and writes to obj/(debug|release)/package/packagetmp.
This effectively ignores whether the file is included in the project or
not as well as its build action. It would be better to use as the source
files that are copied to obj/(debug|release)/package/packagetmp by WPP,
since WPP takes into account build action etc.